% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_scm.R
\name{a_scm}
\alias{a_scm}
\title{Augmented Synthetic Control Method for Causal Effect Estimation}
\usage{
a_scm(df, params, cpd = FALSE, window = 10)
}
\arguments{
\item{df}{Data frame containing the raw data.}

\item{params}{List of parameters. Contains the treatment_group and start_time.}
}
\value{
Returns a list of dataframes with the estimated effects.
}
\description{
The function estimates the causal effect of an intervention using the augmented
synthetic control method. The data is preprocessed, synthetic control methods are applied,
and effects are calculated.
}
\details{
If the `cpd` parameter is set to TRUE, Bayesian structural time series (`bsts`) is
used to compute smoothed residuals. These residuals are then utilized to update the dataframe.

The `augsynth` function is subsequently applied to estimate the causal effect.
}
\examples{
\dontrun{
# Assuming df is your data frame and params is your parameters list:
results <- a_scm(df, params)
}

}
\author{
[Yuqing Dai]
}
