% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date_cols.R
\name{add_date_cols}
\alias{add_date_cols}
\title{Add Date-Based Columns to a Data Frame}
\usage{
add_date_cols(df, date_col, stats)
}
\arguments{
\item{df}{A data frame.}

\item{date_col}{A string specifying the name of the date column in df.}

\item{stats}{A character vector specifying which date-related statistics to add.
Valid options are: "year", "month", "day", "hour", "doy" (day of year),
"dow" (day of week), "unixtime" (UNIX timestamp), "trend", and "all" (to include all statistics).}
}
\value{
A data frame enhanced with the specified date-related columns.
}
\description{
This function enriches a data frame by adding new columns derived from a specified date column.
The new columns can represent various aspects of the date such as year, month, day, hour, etc.
}
\examples{
\dontrun{
df <- data.frame(date = as.POSIXct(c('2021-01-01 14:00', '2021-01-02 15:00')), value = 1:2)
enriched_df <- add_date_cols(df, "date", c("year", "month", "day"))
}

}
\author{
[Yuqing Dai]
}
