% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoMod.R
\name{autoMod}
\alias{autoMod}
\title{Automated Modeling Function}
\usage{
autoMod(
  df,
  response_variable = NULL,
  predictor_variables = "default",
  excluded_variables = NULL,
  max_runtime_secs = NULL,
  max_models = 10,
  split_proportion = 0.8,
  split_by_time = FALSE,
  algorithm = "gbm",
  criterion = "AUTO",
  seed = 1234,
  ...
)
}
\arguments{
\item{df}{Data frame containing the raw data.}

\item{response_variable}{Name of the response (dependent) variable. Default is NULL.}

\item{predictor_variables}{Character vector specifying the predictor (independent) variables. Default is "default".}

\item{excluded_variables}{Variables to be excluded from the modeling process. Default is NULL.}

\item{max_runtime_secs}{Maximum runtime in seconds for the AutoML process. Default is NULL.}

\item{max_models}{Maximum number of models for the AutoML process. Default is 10.}

\item{split_proportion}{Proportion of data to use for training. Default is 0.8.}

\item{split_by_time}{Logical indicating whether to split data based on time. Default is FALSE.}

\item{algorithm}{Algorithm type for best model extraction. Default is "gbm".}

\item{criterion}{Criterion for best model extraction. Default is "AUTO".}

\item{seed}{Seed for reproducibility. Default is 1234.}

\item{...}{Other arguments passed to the function.}
}
\value{
Returns a list containing the final model, training data, testing data, predictions, and the leaderboard.
}
\description{
This function provides an automated workflow for modeling data using H2O's AutoML capabilities.
}
\examples{

\dontrun{
set.seed(123)
df <- data.frame(
  datetime = seq(as.POSIXct("2020-01-01 00:00:00"), as.POSIXct("2020-01-10 00:00:00"), by = "hour"),
  A = sample(c(0:100), size = 217, replace = TRUE),
  B = sample(c(0:100), size = 217, replace = TRUE),
  C = sample(c(0:100), size = 217, replace = TRUE),
  D = sample(c(0:100), size = 217, replace = TRUE),
  E = sample(c(0:100), size = 217, replace = TRUE),
  G = sample(c(0:100), size = 217, replace = TRUE))
#'
#' Randomly replace some values with NA
df$A[sample(1:217, 50)] <- NA
df$B[sample(1:217, 30)] <- NA
df$C[sample(1:217, 20)] <- NA
#'
df
#'
aml_test <- autoMod(df, response_variable = "A", predictor_variables = c("E","G"), max_models = 10, split_proportion = 0.8, seed = 123, algorithm = "gbm")
aml_test[[5]]
#'
pred <- h2o.predict(object = aml_test[[1]], newdata = aml_test[[3]])
plot(as.data.frame(pred)$predict, as.data.frame(aml_test[[3]])$y)

exa <- h2o.explain(aml_test[[1]], aml_test[[3]])
exa
}

}
\author{
[Yuqing Dai]
}
