% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_change_points.R
\name{detect_change_points}
\alias{detect_change_points}
\title{Detect Change Points in Time Series Data}
\usage{
detect_change_points(data, start_time, end_time)
}
\arguments{
\item{data}{A data frame containing time series data.}

\item{start_time}{A character or Date object specifying the start time for analysis.}

\item{end_time}{A character or Date object specifying the end time for analysis.}
}
\value{
Returns plots showing the detected change points using both methods.
}
\description{
This function detects change points in the time series data using two methods:
the strucchange package and the changepoint package with the PELT method.
}
\examples{
\dontrun{
Create a sample data frame for testing
set.seed(123)
sample_data <- data.frame(
  datetime = seq(as.Date("2023/01/01"), as.Date("2023/06/30"), by = "day"),
  no2 = c(rnorm(45, 11, 1),rnorm(45, 10, 1), rnorm(25, 14, 1), rnorm(20, 12, 1), rnorm(46, 10, 1))
)

Check the dataframe
plot(sample_data)
Call the function with the csv files just created
detect_change_points(sample_data, start_time = "2023-01-01", end_time = "2023-06-30")
}

}
\author{
[Yuqing Dai]
}
