% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2df.R
\name{list2df}
\alias{list2df}
\title{Convert List to Data Frame Based on Selected Column}
\usage{
list2df(df_list, params)
}
\arguments{
\item{df_list}{A list of data frames to be combined.}

\item{params}{A list containing parameters for processing. The list should include a 'response_variable' element.}
}
\value{
Returns a consolidated data frame.
}
\description{
This function processes a list of data frames and combines them into a single data frame.
A specified column name is created by appending a suffix to the response variable provided in the parameters.
}
\examples{
\dontrun{
data1 <- data.frame(a = 1:3, b = 4:6)
data2 <- data.frame(a = 7:9, b = 10:12)
mylist <- list(df1 = data1, df2 = data2)
myparams <- list(response_variable = "a")

combined_df <- list2df(mylist, myparams)
print(combined_df)
}

}
\author{
[Your Name or Your Team's Name]
}
