% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2df_xyselect.R
\name{list2df_xyselect}
\alias{list2df_xyselect}
\title{Combine Selected Columns from List of Data Frames}
\usage{
list2df_xyselect(df_list, colnames = "y", time_resolution = NULL)
}
\arguments{
\item{df_list}{A list of data frames to be combined.}

\item{colnames}{Character vector specifying the column names to extract from each data frame in the list. Default is 'y'.}

\item{time_resolution}{Optional: A time resolution for aggregating data. Valid values might include "minute", "hour", "day", etc.}
}
\value{
Returns a consolidated data frame.
}
\description{
This function processes a list of data frames and consolidates them based on the specified columns.
The data frames in the list are expected to contain a 'datetime' column, and they are merged by this column.
}
\examples{
\dontrun{
data1 <- data.frame(datetime = as.POSIXct(c("2021-01-01", "2021-01-02")), y = c(1,2))
data2 <- data.frame(datetime = as.POSIXct(c("2021-01-01", "2021-01-02")), y = c(3,4))
mylist <- list(df1 = data1, df2 = data2)

combined_df <- list2df_xyselect(mylist, colnames = "y")
print(combined_df)
}

}
\author{
[Yuqing Dai]
}
