% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_na_rm.R
\name{list_na_rm}
\alias{list_na_rm}
\title{Remove Data Frames with Specified Columns All NA}
\usage{
list_na_rm(data_list, col_names)
}
\arguments{
\item{data_list}{A list of data frames to process.}

\item{col_names}{A character vector of column names to check for all NA values.}
}
\value{
Returns a filtered list of data frames.
}
\description{
This function processes a list of data frames and removes any data frame
that contains specified columns with all NA values.
}
\examples{
\dontrun{
data1 <- data.frame(a = c(1, 2, NA), b = c(NA, NA, NA))
data2 <- data.frame(a = c(3, 4, 5), b = c(NA, 1, 2))
mylist <- list(df1 = data1, df2 = data2)

filtered_list <- list_na_rm(mylist, col_names = "b")
print(filtered_list)
}

}
\author{
[Yuqing Dai]
}
