% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_envir.R
\name{load_envir}
\alias{load_envir}
\title{Load Environment for Analysis}
\usage{
load_envir(mirror_url = "https://cran.ma.imperial.ac.uk/")
}
\arguments{
\item{mirror_url}{A character string specifying the preferred CRAN mirror URL.}
}
\description{
This function installs and loads the necessary packages.
It provides the user with an option to select their preferred CRAN mirror for package installations.
}
\examples{
\dontrun{
load_envir(mirror_url = "https://cran.ma.imperial.ac.uk/")
}

}
\author{
[Yuqing Dai]
}
