% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_dfs.R
\name{merge_dfs}
\alias{merge_dfs}
\title{Merge a List of Data Frames Based on a Common Column}
\usage{
merge_dfs(df_list, colnames)
}
\arguments{
\item{df_list}{A list of data frames to be merged. Each data frame should contain a 'datetime' column.}

\item{colnames}{Character vector of column names to be considered during the merge.
For these columns, the function will prioritize non-NA values over NA values.}
}
\value{
A merged data frame.
}
\description{
This function takes a list of data frames and merges them based on a common 'datetime' column.
For columns specified in `colnames`, if there are duplicate columns due to the merge, the function will
prioritize non-NA values over NA values.
}
\examples{
\dontrun{
df1 <- data.frame(datetime = c("2021-01-01", "2021-01-02"), x = c(1, NA), y = c(3, 4))
df2 <- data.frame(datetime = c("2021-01-01", "2021-01-02"), x = c(NA, 2), z = c(5, 6))
df_list <- list(df1 = df1, df2 = df2)

merged_df <- merge_dfs(df_list, colnames = c("x"))
print(merged_df)
}

}
\author{
[Yuqing Dai]
}
