% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_scatter_plot.R
\name{missing_scatter_plot}
\alias{missing_scatter_plot}
\title{Scatter Plot Highlighting Missing Data}
\usage{
missing_scatter_plot(
  df,
  x_col,
  y_col,
  title = "Scatter Plot with Missing Data Highlighted",
  color_scheme = c(`Existing data` = "black", `Missing data` = "red")
)
}
\arguments{
\item{df}{A data.frame that contains the data.}

\item{x_col}{The name of the column to be plotted on the x-axis.}

\item{y_col}{The name of the column to be plotted on the y-axis.}

\item{title}{The title of the scatter plot. Default is "Scatter Plot with Missing Data Highlighted".}

\item{color_scheme}{A named vector indicating the colors for highlighting the existing and missing data points.
Default is `c("Existing data" = "black", "Missing data" = "red")`.}
}
\value{
A ggplot scatter plot.
}
\description{
This function creates a scatter plot for two specified columns of a data frame and highlights
the data points that have missing values in either or both of the columns. The plot provides
a visual representation of the distribution of missing and available data points.
}
\examples{
\dontrun{
 df <- data.frame(x = c(1, 2, NA, 4, 5),
                 y = c(NA, 3, 4, 5, NA))

 Create scatter plot with missing data highlighted
 missing_scatter_plot(df, "x", "y", "My Scatter Plot", c("blue", "red"))
}

}
\author{
[Yuqing Dai]
}
