% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_treat.R
\name{missing_treat}
\alias{missing_treat}
\title{Handle Missing Values in a Data Frame Using Various Imputation Techniques}
\usage{
missing_treat(
  df,
  method = "rm",
  response_variable = NULL,
  predictor_variables = NULL,
  split_by_time = F,
  split_proportion = 0.8,
  seed = 1234,
  patch_data = T,
  max_models = 10,
  algorithm = "gbm",
  criterion = "AUTO",
  max_runtime_secs = NULL,
  ...
)
}
\arguments{
\item{df}{A data.frame with potential missing values.}

\item{method}{The method to handle missing values. Default is "rm" (remove rows with missing values).
It can take values like "h2o", "linear", "mean", "median", etc.}

\item{response_variable}{The target/response variable(s) for imputation. Default is NULL.}

\item{predictor_variables}{Predictor variables for the imputation. Default is NULL.}

\item{split_by_time}{Logical. If TRUE, data splitting considers time-based order. Default is FALSE.}

\item{split_proportion}{Proportion of data to be used for training in case of H2O imputation. Default is 0.8.}

\item{seed}{Seed for reproducibility. Default is 1234.}

\item{patch_data}{Logical. If TRUE, replaces the missing values in the original dataframe with the predicted values. Default is TRUE.}

\item{max_models}{Maximum number of models to be trained using H2O's autoML. Default is 10.}

\item{algorithm}{Algorithm choice for H2O's autoML imputation. Default is "gbm".}

\item{criterion}{Criterion choice for H2O's autoML imputation. Default is "AUTO".}

\item{max_runtime_secs}{Maximum time (in seconds) to run H2O's autoML. If NULL, there's no limit. Default is NULL.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A data frame with imputed values.
}
\description{
This function provides a versatile approach to handling missing data in a given data frame. Depending on the method chosen,
it can remove rows with missing values, use linear interpolation, replace with mean/median, or utilize the sophisticated
capabilities of H2O's autoML for imputation.
}
\examples{
\dontrun{
set.seed(123)
df <- data.frame(
  datetime = seq(as.POSIXct("2020-01-01 00:00:00"), as.POSIXct("2020-01-10 00:00:00"), by = "hour"),
  A = sample(c(0:100), size = 217, replace = TRUE),
  B = sample(c(0:100), size = 217, replace = TRUE),
  C = sample(c(0:100), size = 217, replace = TRUE),
  D = sample(c(0:100), size = 217, replace = TRUE),
  E = sample(c(0:100), size = 217, replace = TRUE),
  G = sample(c(0:100), size = 217, replace = TRUE))

Randomly replace some values with NA
df$A[sample(1:217, 50)] <- NA
df$B[sample(1:217, 30)] <- NA
df$C[sample(1:217, 20)] <- NA

predictor_variables <- c("D","E","G")

test <- missing_treat(df, method = "h2o", response_variable = c("A", "B"), predictor_variables = c("D","E","G"), max_runtime_secs = 5, seed = 123, patch_data = T, split_proportion = 0.8)
test1 <- missing_treat(df, method = "mean", response_variable = c("A", "B"))
test2 <- missing_treat(df, method = "rm", response_variable = c("A", "B"))
response_variable <- c("A","B")
}

}
\author{
[Yuqing Dai]
}
