% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_stats.R
\name{mod_stats}
\alias{mod_stats}
\title{Evaluate the performance of a model compared to observed data.}
\usage{
mod_stats(
  mydata,
  mod = "mod",
  obs = "obs",
  stats = c("n", "FB", "MG", "FAC2", "VG", "r", "RMSEs", "RMSEu", "RMSE", "COE", "IOA",
    "IOAr", "default")
)
}
\description{
Evaluate the performance of a model compared to observed data.
}
