% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mlscm.R
\name{plot_mlscm}
\alias{plot_mlscm}
\title{Multi-Resolution Time Series Plot}
\usage{
plot_mlscm(
  df,
  y_variable = NULL,
  highlight = NULL,
  x_range = range(df$datetime, na.rm = TRUE),
  start_times = NULL,
  end_times = NULL,
  time_resolution = "1 hour",
  se = FALSE,
  plotly = F,
  span = 0.75,
  ylab = "Value",
  xlab = "Datetime",
  title = "Timeseries of treatment and control elements",
  subtitle = "",
  legend = "right",
  text_size = 1,
  line_size = 1,
  color = "gray90",
  ylim = NULL,
  line_colors = NULL,
  ...
)
}
\arguments{
\item{df}{A data frame containing the data to be plotted, including a datetime column.}

\item{y_variable}{Character vector specifying which columns in `df` are to be used as y-variables.
If NULL, all columns except 'datetime' will be used. Default is NULL.}

\item{highlight}{Character vector specifying which y-variables should be highlighted. Default is NULL.}

\item{x_range}{A vector of two datetimes specifying the start and end of the x-axis.
Default is the range of the datetime column.}

\item{start_times}{A vector of start times to indicate special periods in the plot. Default is NULL.}

\item{end_times}{A vector of end times corresponding to `start_times` to indicate special periods in the plot. Default is NULL.}

\item{time_resolution}{Character string indicating the granularity for data aggregation. Default is "1 hour".}

\item{se}{Logical indicating if the standard error bands should be plotted around the smooth line. Default is FALSE.}

\item{plotly}{Logical indicating if the output should be a plotly plot. Default is FALSE.}

\item{span}{Numeric controlling the amount of smoothing for loess method. Default is 0.75.}

\item{ylab}{Character string for y-axis label. Default is "Value".}

\item{xlab}{Character string for x-axis label. Default is "Datetime".}

\item{title}{Character string for the main plot title. Default is "Timeseries of treatment and control elements".}

\item{subtitle}{Character string for the plot subtitle. Default is an empty string.}

\item{legend}{Position of the legend. Default is "right".}

\item{text_size}{Numeric controlling the size of text elements in the plot. Default is 1.0.}

\item{line_size}{Numeric controlling the size of lines in the plot. Default is 1.0.}

\item{color}{Color for the lines. Default is "gray90".}

\item{ylim}{Numeric vector of length 2 indicating y-axis limits. Default is NULL.}

\item{line_colors}{Vector of colors for lines if custom colors are required. Default is NULL.}

\item{...}{Further arguments passed to other methods.}
}
\value{
A ggplot or plotly object, invisibly.
}
\description{
This function plots a time series based on provided data, offering multiple visual and analytical features.
}
\details{
The function takes in a dataframe with a datetime column and plots it as a time series.
Users can specify multiple y-variables, highlight certain variables, indicate special periods with `start_times` and `end_times`,
aggregate data at different time resolutions, and add smooth curves with optional standard error bands.
The resulting plot can be either a ggplot or a plotly interactive plot.
}
\examples{
\dontrun{
df <- data.frame(
  datetime = seq(from = as.POSIXct("2020-01-01 00:00"), to = as.POSIXct("2020-01-10 23:00"), by = "hour"),
  variable1 = rnorm(240),
  variable2 = rnorm(240)
)

plot_mlscm(df, y_variable = c("variable1", "variable2"), highlight = "variable1", start_times = c("2020-01-03 00:00"), end_times = c("2020-01-07 23:00"))
}

}
\author{
[Yuqing Dai]
}
