% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsvat2.R
\name{plot_tsvat2}
\alias{plot_tsvat2}
\title{Plot Timeseries of Treatment and Control Elements}
\usage{
plot_tsvat2(
  df,
  y_variable = NULL,
  highlight = NULL,
  x_range = NULL,
  start_time = NULL,
  end_time = NULL,
  buffer_time = NULL,
  time_resolution = "1 hour",
  plotly = FALSE,
  ylab = "Value",
  xlab = "Date",
  title = "Timeseries of treatment and control elements",
  subtitle = "",
  legend.pos = "right",
  text_size = 1,
  line_size = 1,
  ylim = NULL,
  layer_order = NULL,
  line_colors = NULL,
  ...
)
}
\arguments{
\item{df}{A data frame with a 'datetime' column and one or more columns of numeric data.}

\item{y_variable}{A character vector specifying which columns to plot. Default is all columns except 'datetime'.}

\item{highlight}{A character vector of column names that should be highlighted in the plot.}

\item{x_range}{A vector specifying the start and end datetime to be plotted.}

\item{start_time}{A datetime indicating when a vertical line should be added, typically representing the start of an event.}

\item{end_time}{A datetime indicating when a vertical line should be added, typically representing the end of an event.}

\item{buffer_time}{A datetime indicating when a vertical line should be added, often representing a buffer or lag.}

\item{time_resolution}{A character string that represents the frequency to aggregate the data. Default is "1 hour".}

\item{plotly}{A logical. If TRUE, returns an interactive Plotly plot. Otherwise, returns a ggplot. Default is FALSE.}

\item{ylab}{A label for the y-axis. Default is "Value".}

\item{xlab}{A label for the x-axis. Default is "Date".}

\item{title}{The main title for the plot.}

\item{subtitle}{A subtitle for the plot.}

\item{legend.pos}{The position of the legend on the plot. Options are "right", "left", "top", "bottom", etc.}

\item{text_size}{A numeric value to scale the size of text elements in the plot. Default is 1.0.}

\item{line_size}{A numeric value to scale the size of line elements in the plot. Default is 1.0.}

\item{ylim}{A numeric vector of length 2 indicating the y limits.}

\item{layer_order}{A character vector specifying the order in which to layer the plotted series.}

\item{line_colors}{A named list or vector of colors for the lines.}

\item{...}{Further arguments passed to other methods.}
}
\value{
If `plotly` is TRUE, returns a Plotly object. Otherwise, returns a ggplot object.
}
\description{
This function visualizes a timeseries of data in the provided dataframe. It supports highlighting certain series,
specifying a date range, and various formatting options. Additionally, the resulting plot can either be
in ggplot2 format or converted to a Plotly interactive plot.
}
\details{
The function offers a flexible way to visualize timeseries data with options to highlight certain series,
add vertical lines to indicate specific events, and customize the appearance of the plot.
}
\examples{
\dontrun{
data <- data.frame(datetime = as.POSIXct(c("2022-01-01", "2022-01-02", "2022-01-03", "2022-01-04", "2022-01-05")),
                   value1 = c(1, 2, 3, 4, 5),
                   value2 = c(5, 4, 3, 2, 1))
plot_tsvat2(data, y_variable = c("value1", "value2"), highlight = "value2")
}

}
\author{
[Yuqing Dai]
}
