% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read Data From Specified Directory and Process It}
\usage{
read_data(params, write_out = FALSE, ...)
}
\arguments{
\item{params}{A named list containing various parameters:
- `data_dir`: The directory containing data files.
- `file_pattern`: A regex pattern to match specific files within the directory.
- `datetime_format`: The format of the datetime column in the files.
- `data_timerange`: A vector of two Date elements specifying the range of dates to filter the data.
- `dependent_variable`: The name of the dependent variable in the dataset.
- `wenormed`: A logical value indicating whether to modify the dependent variable name with a "_wn" suffix.}

\item{write_out}{A logical. If TRUE, writes the processed data to a file named "df_origin.csv".
Defaults to FALSE.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A list containing:
  - `list_df`: A named list of data frames, where each data frame corresponds to a file and
    the name is the filename without the extension.
  - `df_origin`: A combined data frame derived from `list_df`.
}
\description{
This function reads multiple data files from a specified directory, processes them,
and optionally writes the processed data to a CSV file. The function supports both
CSV and XLSX file formats.
}
\details{
The function first gets a list of filenames matching the given pattern in the directory.
It then reads each file, processes the datetime column, filters rows based on the given date range,
adds additional date-related columns, and ensures that the dependent variable is numeric.
If the `deweathered` parameter is TRUE, it modifies the dependent variable name by adding a "_wn" suffix.
Finally, it optionally writes the processed data to a CSV file if `write_out` is TRUE.
}
\examples{
\dontrun{
params <- list(
    data_dir = "path/to/data/",
    file_pattern = "*.csv",
    datetime_format = "\%Y-\%m-\%d \%H:\%M:\%S",
    data_timerange = as.Date(c("2022-01-01", "2022-12-31")),
    dependent_variable = "target",
    deweathered = FALSE
)

result <- read_data(params)
head(result$df_origin)
}

}
\author{
[Yuqing Dai]
}
