% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rds.R
\name{read_rds}
\alias{read_rds}
\title{Load Multiple RDS Files into a Named List}
\usage{
read_rds(params)
}
\arguments{
\item{params}{A list that contains configuration parameters for reading the `.rds` files.
It should include the following elements:
\itemize{
  \item `data_dir`: A character string specifying the directory where the `.rds` files are located.
}}
}
\value{
A named list where each item contains the content of one `.rds` file.
The names of the list items correspond to the base filenames of the `.rds` files (without the extension).
}
\description{
This function reads all `.rds` files located in a specified directory into a named list,
where each list item corresponds to the content of one `.rds` file.
}
\details{
The function first corrects the directory separator to ensure compatibility across different platforms.
It then lists all the `.rds` files present in the specified directory.
Each `.rds` file is read and stored as an item in a list, with the names of the list items corresponding
to the base filenames of the `.rds` files (without the extension).
}
\examples{
\dontrun{
# Assume there are two .rds files in the "data/" directory: "file1.rds" and "file2.rds".
params <- list(data_dir = "data/")
result <- read_rds(params)
print(names(result))  # Should output: "file1" "file2"
}

}
\seealso{
\code{\link[base]{readRDS}}, \code{\link[tools]{file_path_sans_ext}}
}
\author{
[Yuqing Dai]
}
