% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_column.R
\name{remove_columns}
\alias{remove_columns}
\title{Remove Specified Columns from a Data Frame}
\usage{
remove_columns(df, cols_to_remove)
}
\arguments{
\item{df}{A data frame from which columns will be removed.}

\item{cols_to_remove}{A character vector specifying the names of the columns to be removed from the data frame.}
}
\value{
A data frame with the specified columns removed.
}
\description{
This function takes in a data frame and a vector of column names, and returns the data frame with the specified columns removed.
}
\details{
The function leverages the `select` function from the `dplyr` package to safely deselect columns from the input data frame.
It uses the `all_of` function to handle column names provided in the `cols_to_remove` parameter, ensuring that if any of the
specified columns do not exist in the data frame, no error is thrown.
}
\examples{
\dontrun{
data <- data.frame(
  A = 1:5,
  B = 6:10,
  C = 11:15
)

# Remove columns A and C
modified_data <- remove_columns(data, c("A", "C"))
print(modified_data)
}

}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{all_of}}
}
\author{
[Yuqing Dai]
}
