% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_plot.R
\name{shadow_plot}
\alias{shadow_plot}
\title{Function to show missing value}
\usage{
shadow_plot(
  df,
  column = NULL,
  date_col = NULL,
  plot_title = "Shadow Plot of Missing Data",
  x_label = "Variables",
  y_label = "Observations",
  na_color = "red",
  non_na_color = "blue",
  na_label = "Missing",
  non_na_label = "Not Missing"
)
}
\arguments{
\item{df}{A dataframe for which to generate the shadow plot.}

\item{column}{A character vector specifying the names of the columns to include in the
plot. If NULL (default), all columns in the dataframe are included.}

\item{date_col}{A character string specifying the name of the date column in the
dataframe. If NULL (default), no date column is used.}

\item{plot_title}{A character string specifying the title of the plot. Default is
"Shadow Plot of Missing Data".}

\item{x_label}{A character string specifying the label for the x-axis. Default is
"Variables".}

\item{y_label}{A character string specifying the label for the y-axis. Default is
"Observations".}

\item{na_color}{A character string specifying the color for missing values. Default is
"red".}

\item{non_na_color}{A character string specifying the color for non-missing values.
Default is "blue".}

\item{na_label}{A character string specifying the label for missing values. Default is
"Missing".}

\item{non_na_label}{A character string specifying the label for non-missing values.
Default is "Not Missing".}
}
\value{
A ggplot object representing the shadow plot.
}
\description{
shadow_plot
}
\details{
This function generates a shadow plot to visualize the missing values in a dataframe.
Each row of the plot corresponds to a row in the dataframe, and each column corresponds
to a variable (column) in the dataframe. Missing values are colored differently from
non-missing values, making it easy to identify patterns of missingness.
}
\note{
This function requires the ggplot2 and dplyr packages. Make sure these packages
are installed before using this function.
}
\author{
[Yuqing Dai]
}
