% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_missing.R
\name{show_missing}
\alias{show_missing}
\title{Visualize Missing Data in a Data Frame}
\usage{
show_missing(
  df,
  column = "all",
  color = "darkorchid",
  date = FALSE,
  type = "NULL",
  threshold = 0,
  sort_by = "percentage",
  show_shadow_plot = FALSE,
  date_col = NULL,
  show_scatter_plot = FALSE,
  x_col = NULL,
  y_col = NULL,
  ...
)
}
\arguments{
\item{df}{A data frame for which missing data insights are to be generated.}

\item{column}{A character vector or "all" specifying which columns to consider.
Default is "all".}

\item{color}{A color string for the histograms. Default is "darkorchid".}

\item{date}{A logical indicating if date columns should be added. Default is FALSE.}

\item{type}{A character string specifying the type of temporal aggregation
(e.g., "year", "month"). Default is "NULL".}

\item{threshold}{A numeric threshold. Only columns with a missing percentage
above this threshold will be displayed. Default is 0.}

\item{sort_by}{A character string indicating the sorting variable for the histogram.
Options are "percentage" or the column name. Default is "percentage".}

\item{show_shadow_plot}{A logical indicating if a shadow plot should be generated.
Default is FALSE.}

\item{date_col}{A character string specifying the date column.
Relevant if date is TRUE.}

\item{show_scatter_plot}{A logical indicating if a scatter plot should be generated.
Default is FALSE.}

\item{x_col}{A character string specifying the x-axis column for the scatter plot.}

\item{y_col}{A character string specifying the y-axis column for the scatter plot.}

\item{...}{Further arguments passed to other methods.}
}
\value{
A named list containing:
\itemize{
  \item `missing_data`: A data frame summarizing missing data by column or time.
  \item `missing_plot`: A ggplot object visualizing the missing data.
  \item `shadow_plot`: (Optional) A ggplot object visualizing the data's "shadow" or
                        absence/presence of data.
  \item `scatter_plot`: (Optional) A ggplot object showing a scatter plot of the data with missing values highlighted.
}
}
\description{
This function provides insights into the distribution and patterns of missing data in a
data frame. It supports creating histograms, shadow plots, and scatter plots.
}
\details{
The function provides various visualizations and summaries of missing data in a given data frame.
Users can request a histogram showing missing data percentages by column or by time (e.g., year, month).
Optionally, users can visualize missing data patterns using shadow plots and scatter plots.
}
\author{
[Yuqing Dai]
}
