% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sum.R
\name{show_sum}
\alias{show_sum}
\title{Generate Summary Statistics for a Data Frame}
\usage{
show_sum(df, date_col = "datetime", interval = "default")
}
\arguments{
\item{df}{A data frame containing the data to be summarized.}

\item{date_col}{A character string specifying the name of the date column in the data frame.
Defaults to "datetime".}

\item{interval}{A character string specifying the time interval for grouping the summary.
Possible values are "default", "year", "month", "day", and "hour". If "default",
statistics are generated without grouping. Defaults to "default".}
}
\value{
A data frame containing the summary statistics. Each row represents a variable, and the columns
include statistics such as min, q1, median, mean, q3, max, and sd for numeric variables,
and n_unique and levels for factor variables.
}
\description{
This function produces summary statistics for numeric and factor columns in a given data frame.
The statistics can be generated for the entire data frame or be grouped by specific time intervals.
}
\details{
The function first validates the input data frame and parameters. It ensures that the provided
date column exists in the data frame and is of an appropriate date-time type.
The numeric columns are summarized with min, 1st quartile, median, mean, 3rd quartile, max, and standard deviation.
Factor columns are summarized with the number of unique values and the levels.

When the interval is specified (other than "default"), the function groups the data by the specified
time interval and computes the summary statistics for each group.
}
\examples{
\dontrun{
Create a sample data frame with missing values
set.seed(12)
 df <- data.frame(
 datetime = seq(as.POSIXct("2020-01-01 00:00:00"), as.POSIXct("2020-01-10 00:00:00"), by = "hour"),
 A = sample(c(0:100), size = 217, replace = TRUE),
 B = sample(c(0:100), size = 217, replace = TRUE),
 C = sample(c(0:100), size = 217, replace = TRUE))

Randomly replace some values with NA
df$A[sample(1:217, 50)] <- NA
df$B[sample(1:217, 30)] <- NA
df$C[sample(1:217, 20)] <- NA

Use the show_missing function to display the missing data distribution, shadow plot, and scatter plot
show_sum(df, interval = "hour")
}

}
\author{
[Yuqing Dai]
}
