% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_policy_time2.R
\name{split_policy_time2}
\alias{split_policy_time2}
\title{Split Data Based on Specified Time Intervals}
\usage{
split_policy_time2(df, start_time = NULL, end_time = NULL)
}
\arguments{
\item{df}{A data frame that contains a datetime column to be used for splitting.}

\item{start_time}{A character string or a POSIXct object specifying the beginning of the event period.}

\item{end_time}{A character string or a POSIXct object specifying the end of the event period.}
}
\value{
A named list containing three data frames (`pre_event`, `event`, `post_event`), each corresponding
to the aforementioned segments.
}
\description{
This function partitions a data frame into three segments (pre-event, event, and post-event)
based on the specified start and end times.
}
\details{
The function first conducts input checks to ensure that the provided data frame contains a `datetime` column
and that both `start_time` and `end_time` are provided. It then splits the data frame into three segments:
\itemize{
  \item `pre_event`: Entries with a datetime before the specified `start_time`.
  \item `event`: Entries within the inclusive range of `start_time` and `end_time`.
  \item `post_event`: Entries with a datetime after the specified `end_time`.
}
All data frames are returned within a named list.
}
\examples{
\dontrun{
data <- data.frame(datetime = as.POSIXct(c("2022-01-01 12:00", "2022-01-02 12:00", "2022-01-03 12:00")),
                   value = c(10, 20, 30))
result <- split_policy_time2(data, start_time = "2022-01-02", end_time = "2022-01-02")
print(result$pre_event)  # Data before 2022-01-02
print(result$event)      # Data on 2022-01-02
print(result$post_event) # Data after 2022-01-02
}

}
\author{
[Yuqing Dai]
}
