% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_aggregation.R
\name{time_aggregation}
\alias{time_aggregation}
\title{Aggregate Data by Specified Time Resolution}
\usage{
time_aggregation(df, time_resolution = NULL, ...)
}
\arguments{
\item{df}{A data.frame: the input data set.}

\item{time_resolution}{A character string indicating the desired time resolution for aggregation (e.g., "day", "hour"). If NULL, no time aggregation is applied.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A data frame with aggregated values based on the specified time resolution.
}
\description{
This function aggregates the input data frame based on a provided time resolution.
It handles special cases where wind speed (`ws`) and wind direction (`wd`)
are present in the data by converting them to u and v components and then
aggregating based on the desired time resolution.
}
\details{
If `ws` and `wd` columns are present, the function first converts these
to u and v wind components. It then aggregates the data based on the given
time resolution. For columns other than `ws` and `wd`, the function calculates
the mean for each time period. For the `ws` and `wd` columns, it calculates
the resultant wind speed and direction.

If `ws` and `wd` columns are not present, the function simply aggregates all columns
based on their mean values for each time period.
}
\examples{
\dontrun{
data <- data.frame(
    datetime = as.POSIXct(seq(1, 100, by = 1), origin = "1970-01-01"),
    A = rnorm(100), ws = rnorm(100, mean = 10), wd = runif(100, min = 0, max = 360)
)

aggregated_data <- time_aggregation(data, time_resolution = "day")
head(aggregated_data)
}

}
\author{
[Yuqing Dai]
}
