% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_select.R
\name{xy_select}
\alias{xy_select}
\title{Select Response and Predictor Variables from a Data Frame}
\usage{
xy_select(df, response_variable, predictor_variables = "default", ...)
}
\arguments{
\item{df}{A data.frame that contains the data.}

\item{response_variable}{A vector of response variable(s) to be selected.}

\item{predictor_variables}{A vector of predictor variable(s) to be selected
or "default" to select the default set of predictors: `unixtime, ws, wd, rh, temp, sp`.
Default is "default".}

\item{...}{Additional arguments (currently not used).}
}
\value{
A named list of data frames. Each data frame corresponds to a
       response variable combined with the selected predictor variables.
       The name of the data frame is prefixed with "df_" followed by the
       response variable name.
}
\description{
This function selects the desired response (dependent) variable(s) and
predictor (independent) variables from a data frame. The function allows for
a default set of predictor variables or user-specified predictors. For each
response variable, a separate data frame is constructed with the selected
predictors and is stored in a named list.
}
\examples{
\dontrun{
data <- data.frame(unixtime = 1:5, ws = rnorm(5), wd = rnorm(5),
                  rh = rnorm(5), temp = rnorm(5), sp = rnorm(5), A = rnorm(5), B = rnorm(5))

result <- xy_select(data, response_variable = c("A", "B"), predictor_variables = "default")
print(result$df_A)
}

}
\author{
[Yuqing Dai]
}
