% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMod.R
\name{buildMod}
\alias{buildMod}
\title{Building Models Using WENORM}
\usage{
buildMod(df, params, start_index = 1, end_index = NULL, ...)
}
\arguments{
\item{df}{A list of data frames to process.}

\item{params}{A list of parameters specifying details like predictor variables, output directory, wenorm method, etc.}

\item{start_index}{The index at which to start processing the list of data frames. Default is 1.}

\item{end_index}{The index at which to end processing the list of data frames. If NULL, it defaults to the length of the input list.}

\item{...}{Other arguments passed to the function.}
}
\value{
Returns a list containing processed data frames (list_df), processed data with XY select (data_wn), baseline data (base_line), and AQMOD data (aqmod).
}
\description{
This function processes and builds models using the `wenorm` method based on provided parameters.
}
\examples{
\dontrun{
# Example usage with dummy data and parameters:
results <- buildMod(df_list, params)
}

}
\author{
[Yuqing Dai]
}
