% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wenorm.R
\name{run_wenorm}
\alias{run_wenorm}
\title{Execute Data Preprocessing and Model Training Using Various Methods}
\usage{
run_wenorm(df, model_params, ...)
}
\arguments{
\item{df}{A data frame that contains the data to be processed and modeled.}

\item{model_params}{A list that provides various model parameters such as the response variable,
predictor variables, method for handling missing data, and various parameters
for the modeling process.}

\item{...}{Further arguments passed to other methods.}
}
\value{
A named list containing:
\itemize{
  \item `df_wenorm`: A data frame resulting from the wenorm processing.
  \item `df_base`: A base data frame resulting from the wenorm processing.
  \item `df_detrend`: A data frame resulting from the detrend wenorm processing.
  \item `df_final`: A data frame resulting from the final wenorm processing.
  \item `aqmod`: A list that contains details about the trained model, including model statistics and a leaderboard.
}
}
\description{
This function preprocesses the provided dataset, handles missing values, and trains a model based on
the method provided in the model parameters. It offers three methods - "aml", "default", and "revised".
}
\details{
The function starts by checking and converting the response and predictor variables to numeric, if necessary.
It then processes the data based on the method provided in `model_params`. The "aml" method processes
and models the data differently than the "default" and "revised" methods. After preprocessing, the function
utilizes the `autoMod` function to train models, followed by the `wenorm` function to further process
the dataset. The final output includes the trained model and processed datasets.
}
\examples{
\dontrun{
sample_data <- data.frame(response = c(1, 2, 3, 4, 5),
                          predictor1 = c(1, NA, 3, 4, 5),
                          predictor2 = c(NA, 2, 3, 4, NA),
                          datetime = as.POSIXct(c("2022-01-01", "2022-01-02", "2022-01-03", "2022-01-04", "2022-01-05")))
params <- list(response_variable = "response",
               predictor_variables = c("predictor1", "predictor2"),
               wenorm_method = "default",
               ...)
result <- run_wenorm(sample_data, model_params = params)
print(result$aqmod$aqmod_stat)
}

}
\author{
[Yuqing Dai]
}
